Modified Generation of Model-Architecture HashTable
This is a bug I encountered by executing the library on my platform. My architecture was not find because of a wrong generation of the Model-Architecture HashTable, which did not allowed to have more models for the same platform, for example:
("GenuineIntel-6-4E-core", SKYLAKE_CORE),
("GenuineIntel-6-5E-core", SKYLAKE_CORE),

This is clearly an error because more model exists of a specific platform.
I therefore realized a dirty fix (but faster): I created a new hashtable dedicated for this in which the (key,value) pair are swapped. 

Hope can help to find something "cleaner".

Cheers