types: adding ioaddr type to paging module
When dealing with DMA-capable devices, we have an additional address
type beyond virtual and physical addresses. This commit adds the IOAddr
as a distinct type. Device drivers use this address to communicate the
location of descriptor queues and buffers to the device.

Currently, this is a copy of the physical address. However, in the
future we might add IOMMU support where the IOAddr will be configurable.

Signed-off-by: Reto Achermann <achreto@gmail.com>